requirements-context.txt
sentence-transformers>=2.7.0
numpy>=1.26.0
tiktoken>=0.7.0
pyyaml>=6.0.1
